+++ packages/botonic-react/src/components/message/message-footer.tsx
@@ -2,7 +2,6 @@ import { useContext } from 'react'
 
 import { SENDERS } from '../../index-types'
 import { WebchatContext } from '../../webchat/context'
-import { MessageFeedback } from './message-feedback'
 import { MessageFooterContainer } from './styles'
 import { MessageTimestamp, resolveMessageTimestamps } from './timestamps'
 
@@ -10,18 +9,12 @@ interface MessageFooterProps {
   enabletimestamps: boolean
   messageJSON: any
   sentBy: SENDERS
-  feedbackEnabled: boolean
-  inferenceId?: string
-  botInteractionId?: string
 }
 
 export const MessageFooter = ({
   enabletimestamps,
   messageJSON,
   sentBy,
-  feedbackEnabled,
-  inferenceId,
-  botInteractionId,
 }: MessageFooterProps) => {
   const { getThemeProperty } = useContext(WebchatContext)
 
@@ -44,13 +37,6 @@ export const MessageFooter = ({
           sentBy={sentBy}
           style={timestampStyle}
           timestamp={messageJSON.timestamp}
-        />
-      ) : null}
-      {feedbackEnabled ? (
-        <MessageFeedback
-          inferenceId={inferenceId}
-          messageId={messageJSON.id}
-          botInteractionId={botInteractionId}
         />
       ) : null}
     </MessageFooterContainer>