+++ packages/botonic-plugin-hubtype-analytics/CHANGELOG.md
@@ -11,11 +11,13 @@ All notable changes to Botonic will be documented in this file.
   </summary>
   
   
-## [0.52.x] - 2026-mm-dd
+## [0.55.x] - 2026-mm-dd
 
 ### Added
 
 ### Changed
+
+- [PR-3258](https://github.com/hubtype/botonic/pull/3258): Remove knowledge base event. 
 
 ### Fixed