+++ packages/botonic-react/CHANGELOG.md
@@ -10,11 +10,13 @@ All notable changes to Botonic will be documented in this file.
     Click to see more.
   </summary>
 
-## [0.53.x] - 2026-mm-dd
+## [0.55.x] - 2026-mm-dd
 
 ### Added
 
 ### Changed
+
+- [PR-3258](https://github.com/hubtype/botonic/pull/3258): Remove system debug trace for knowledge base event. 
 
 ### Fixed